Working prototype

+++ main.py
@@ -8,7 +8,8 @@ Created on Tue Apr 21 20:49:07 2020
 import platform
 import time
 from datetime import datetime as dt
-from pathlib.Path import resolve
+from pathlib import Path
+from os import path
 
 # for initially windows only
 if (platform.release() != "Windows"):
@@ -19,7 +20,7 @@ if (platform.release() != "Windows"):
 # for windows 10 x64
 # host_file = "C:\Windows\System32\drivers\etc\hosts"
 host_file = "hosts"
-websites_file = "websites.txt"
+websites_file = "blocklist.txt"
 backup_host_file = "hosts.backup"
 
 redirect = "0.0.0.0"    # I can modify this to
@@ -31,27 +32,25 @@ create backup of original host file.
 """
 
 
-def backup_host_file():
+def backup_host():
     # open host file and read it's content
     with open(host_file, "r", encoding="utf-8") as f:
         host_file_content = f.read()
 
     with open(backup_host_file, "w", encoding="utf-8") as f:
-        f.write(host_file_content)
+        f.write(str(host_file_content))
 
 
-def restore_host_file():
-    try:
-        backup_host_file.resolve(strict=True)
-    except FileNotFoundError:
-        print("\nNo backup exists")
-    else:
-        print("\nBackup File Exists. Restoring!!")
+def restore_host():
+    if(path.exists(backup_host_file)):
+        print("\nTrying to restore!!")
         with open(backup_host_file, "r", encoding="utf-8") as f:
             backup_host_content = f.read()
         with open(host_file, "w", encoding="utf-8") as f:
             f.write(backup_host_content)
         print("\nRestored the host file from host backup successfully!")
+    else:
+        print("Backup Not Found!")
 
 
 def add_block_sites():
@@ -77,12 +76,14 @@ def time_manager(start_time, end_time):
 
 def main():
     print("I'm OAT!")
-    start_time = int(input("Enter the start time in 24 Hours format.\nExample\t2 for 2 AM 18 for 6PM.))
+    start_time = int(input("Enter the start time in 24 Hours format.\nExample\t2 for 2 AM 18 for 6PM."))
 
-    end_time = int(input("Similarly, enter end time!))
-    if((0<= start_time <=24) && (0<= endtime <= 24 )):
-        backup_host_file()
-        time_mnager(start_time, end_time)
+    end_time = int(input("Similarly, enter end time!"))
+    if(((0<= start_time <=24) and (0<= end_time <= 24 )) == True):
+        backup_host()
+        time_manager(start_time, end_time)
+        restore_host()
+    exit()
 
 
 if __name__ == "__main__":
